--- axios/upload.axios.ts
@@ -1,10 +1,0 @@
-export default {}
-
-// export const config = {
-//     headers: { 'content-type': 'multipart/form-data' },
-//     onUploadProgress: (event) => {
-//         console.log(`Current progress:`, Math.round((event.loaded * 100) / event.total));
-//     },
-// };
-
-// export const response = await axios.post('/api/uploads', formData, config);
