+++ const/const.ts
@@ -1,0 +1,2 @@
+export const UPLODED_FILE_DESTINATION_PATH = "./public/uploads"
+export const UPLOAD_FILE_FORM_FIELD_NAME = "upload_file_form_field"
