--- types/Positon.ts
@@ -1,6 +1,0 @@
-// sample
-export type Position = {
-    x: number,
-    y: number,
-    z: number
-}
